NEX-15: Reorganize component directory
* add directories for media, paragraph and node components
