+++ next/components/node/node--page.tsx
@@ -1,4 +1,4 @@
-import { Paragraph } from "@/components/paragraph";
+import { Paragraph } from "@/components/paragraph/paragraph";
 import { PageType, ParagraphType } from "@/types/graphql";
 
 export function NodePage({ page }: { page: PageType }) {
